dbio_read_int*/num/float checks sizes
We need to be clear about what we expect to see in the db file.
Accepting out-of-range integers leads to silent corruption of data.

This also fixes the types/sizes of various object/verb/property flags
and preposition slots as (u)int16_t.  Task ids are int32_t.

Some new rules:

Do not try to make use of [INTMAX_MAX+1..UINTMAX_MAX]; all integers we
care about are in the intmax_t range.  If something needs to be
unsigned, fine; it's in [0..INTMAX_MAX]

Always write out intmax_t (if something somehow got too big, we can
sort it out later -- be lenient about what we write; be strict about
what we read; if you're not running the correct server for a given db
file, that's a problem you need to fix FIRST.)

Do all integer range checks in the intmax_t world to avoid surprises.

+++ db_private.h
@@ -33,8 +33,8 @@ struct Verbdef {
     const char *name;
     Program *program;
     Objid owner;
-    short perms;
-    short prep;
+    uint16_t perms;
+    int16_t  prep;
     Verbdef *next;
 };
 
@@ -55,7 +55,7 @@ struct Proplist {
 typedef struct Pval {
     Var var;
     Objid owner;
-    short perms;
+    uint16_t perms;
 } Pval;
 
 typedef struct Object {
@@ -71,7 +71,7 @@ typedef struct Object {
 
 
     const char *name;
-    int flags;
+    uint16_t flags;
 
     Verbdef *verbdefs;
     Proplist propdefs;
